@bingyanghuang 2019-07-17
This PR is to fix #17654