@luotao1 2019-07-18
Do you have any benchmark result for this mkldnn op on face model in #17654