@jczaja 2019-07-18
@luotao1 When we wanted to test performance we noticed that face model no longer works on develop for us #18658 . So until we have demark model fixed we cannot test performance of leaky rely of mkl-dnn