@luotao1 2019-07-19
Thanks very much for your contribution.
I test it on demark model. leaky_relu speedup from 132ms->38ms on 4 threads, **3.47x speedup**!
- before:
```
Event                        Calls       Total       Min.        Max.        Ave.        Ratio.
thread0::conv2d_transpose    400         3794.74     2.86285     17.8366     9.48685     0.768815
thread0::conv2d              400         899.1       0.729642    3.6068      2.24775     0.182158
thread0::leaky_relu          400         132.44      0.085567    1.10562     0.3311      0.0268324
thread0::batch_norm          300         50.5053     0.1174      0.256459    0.168351    0.0102324
thread0::relu                300         39.7742     0.067462    0.242046    0.132581    0.00805825
thread0::fetch               100         11.2578     0.106524    0.123902    0.112578    0.00228083
thread0::tanh                100         7.4732      0.069471    0.090934    0.074732    0.00151407
thread0::feed                100         0.541868    0.004703    0.006246    0.00541868  0.000109783
```
- after:
```
Event                        Calls       Total       Min.        Max.        Ave.        Ratio.
thread0::conv2d_transpose    400         3809.5      2.86721     17.8962     9.52376     0.786597
thread0::conv2d              400         882.729     0.732503    3.49819     2.20682     0.182269
thread0::batch_norm          300         50.5289     0.119364    0.270264    0.16843     0.0104334
thread0::relu                300         42.0752     0.074177    0.234958    0.140251    0.00868782
thread0::leaky_relu          400         38.5553     0.047176    0.194089    0.0963881   0.007961
thread0::fetch               100         10.8161     0.099839    0.117321    0.108161    0.00223335
thread0::tanh                100         8.28066     0.078006    0.108433    0.0828066   0.00170981
thread0::feed                100         0.526761    0.004705    0.006533    0.00526761  0.000108767
```