@luotao1 2019-07-19
https://github.com/PaddlePaddle/Paddle/blob/067ed70f2def19263d33ef792323c439fa482484/paddle/fluid/framework/ir/is_test_pass.cc#L27-L35
Do you need to update the `op_list` in `is_test_pass`?